pyproject.toml
[tool.poetry]
name = "pollify"
version = "0.0.1"
description = "Service for interactive polls. Web server on flask."
authors = ["Igor Tolmachov <me@igorek.dev>"]
license = "AGPL-3.0-only"
readme = "README.md"

homepage = "https://pollify.igorek.dev"
repository = "https://github.com/ThePollify/server"


[tool.poetry.dependencies]
python = "^3.10"

fastapi = "^0.95.0"
uvicorn = "^0.21.1"
pydantic = "^1.10.7"

SQLAlchemy = { extras = ["asyncio", "mypy"], version = "^2.0.7" }
PyJWT = "^2.6.0"
asyncpg = "^0.27.0"
websockets = "^11.0.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.1.1"
black = "^23.3.0"
isort = "^5.12.0"

[tool.mypy]
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_untyped_calls = true

[tool.isort]
profile = "black"
filter_files = true
skip = ["__init__.py"]
